backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# LangGraph (orchestration only — no LangChain LLM wrappers)
langgraph==0.2.60

# LLM Providers (install all — only the configured one is used at runtime)
google-genai>=1.0.0
openai>=1.30.0
anthropic>=0.30.0

# GitHub
PyGithub==2.5.0
GitPython==3.1.44

# Code Parsing
tree-sitter==0.23.2

# MCP (Model Context Protocol) — server + client SDK
mcp[cli]

# Semgrep (static analysis, called via MCP)
semgrep

# Utilities
python-dotenv==1.0.1
httpx>=0.27.0
python-multipart==0.0.20
